Changed the qtyInputValue to be of type Double to allow for more accurate nutrient calculations, and updated the associated method in search controller to reflect this change.

+++ Nutrivurv/Nutrivurv/LSLSearchController.swift
@@ -88,7 +88,7 @@ class LSLSearchController {
     
     
     
-    func searchForNutrients(qty: Int, measure: String, foodId: String, completion: @escaping (Nutrients?) -> Void) {
+    func searchForNutrients(qty: Double, measure: String, foodId: String, completion: @escaping (Nutrients?) -> Void) {
         let json: [String: Any] = ["ingredients": [["quantity": qty, "measureURI": measure, "foodId": foodId]]]
         let jsonData = try? JSONSerialization.data(withJSONObject: json)
